# example/openapi-example.yaml
openapi: 3.0.3
info:
  title: NocoDB Middleware Example - Books API
  description: |
    A practical example API for managing books, authors, and users with NocoDB Middleware.
    
    This API demonstrates:
    - JWT-based authentication
    - Role-based access control (RBAC)
    - CRUD operations with NocoDB
    - Pagination and filtering
    - Relationships between entities
  version: 1.0.0
  contact:
    name: stritti
    url: https://github.com/stritti/nocodb-middleware
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Local development server
  - url: https://api.example.com/api
    description: Production server
...
paths:
...
  /books:
    get:
      tags:
        - Books
      summary: Get all books
      description: Returns a paginated list of all books with optional filtering
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Book'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Books
      summary: Create a new book
      description: Creates a new book (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookRequest'
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /books/{id}:
    get:
      tags:
        - Books
      summary: Get book by ID
      description: Returns a specific book by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Book details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Books
      summary: Update book
      description: Updates a specific book (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBookRequest'
      responses:
        '200':
          description: Book updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Books
      summary: Delete book
      description: Deletes a specific book (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Book deleted
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /books/search:
    get:
      tags:
        - Books
      summary: Search books
      description: Searches books by title or description
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          description: Search query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Book'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /books/author/{authorId}:
    get:
      tags:
        - Books
      summary: Get books by author
      description: Returns all books by a specific author
      security:
        - bearerAuth: []
      parameters:
        - name: authorId
          in: path
          description: Author ID
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Books by author
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Book'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /books/price-range:
    get:
      tags:
        - Books
      summary: Get books by price range
      description: Returns books within a specific price range
      security:
        - bearerAuth: []
      parameters:
        - name: min
          in: query
          description: Minimum price
          required: true
          schema:
            type: number
            format: float
            minimum: 0
        - name: max
          in: query
          description: Maximum price
          required: true
          schema:
            type: number
            format: float
            minimum: 0
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Books in price range
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Book'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /authors:
    get:
      tags:
        - Authors
      summary: Get all authors
      description: Returns a paginated list of all authors
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
        - $ref: '#/components/parameters/SearchParam'
      responses:
        '200':
          description: List of authors
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Author'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Authors
      summary: Create a new author
      description: Creates a new author (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthorRequest'
      responses:
        '201':
          description: Author created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /authors/{id}:
    get:
      tags:
        - Authors
      summary: Get author by ID
      description: Returns a specific author by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Author details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Author not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Authors
      summary: Update author
      description: Updates a specific author (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAuthorRequest'
      responses:
        '200':
          description: Author updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Author not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Authors
      summary: Delete author
      description: Deletes a specific author (admin only)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Author deleted
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Author not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /authors/{id}/books:
    get:
      tags:
        - Authors
      summary: Get books by author
      description: Returns all books by a specific author
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          description: Author ID
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Books by author
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Book'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /authors/search:
    get:
      tags:
        - Authors
      summary: Search authors
      description: Searches authors by name or bio
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          description: Search query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Author'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login
  parameters:
    IdParam:
      name: id
      in: path
      description: ID of the resource
      required: true
      schema:
        type: integer
    PageParam:
      name: page
      in: query
      description: Page number
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    SortByParam:
      name: sortBy
      in: query
      description: Field to sort by
      required: false
      schema:
        type: string
        default: 'id'
    SortOrderParam:
      name: sortOrder
      in: query
      description: Sort order
      required: false
      schema:
        type: string
        enum: [ASC, DESC]
        default: ASC
    SearchParam:
      name: search
      in: query
      description: Search query
      required: false
      schema:
        type: string
  schemas:
...
    Book:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: "Harry Potter and the Philosopher's Stone"
        description:
          type: string
          example: "The first novel in the Harry Potter series"
        published_year:
          type: integer
          example: 1997
        isbn:
          type: string
          example: "9780747532743"
        price:
          type: number
          format: float
          example: 12.99
        author_id:
          type: integer
          example: 1
        author:
          $ref: '#/components/schemas/Author'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateBookRequest:
      type: object
      required:
        - title
        - price
      properties:
        title:
          type: string
          example: "New Book"
        description:
          type: string
        published_year:
          type: integer
        isbn:
          type: string
        price:
          type: number
          format: float
        author_id:
          type: integer
    UpdateBookRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        published_year:
          type: integer
        isbn:
          type: string
        price:
          type: number
          format: float
        author_id:
          type: integer

    # Authors
    Author:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "J.K. Rowling"
        bio:
          type: string
          example: "British author best known for the Harry Potter series"
        birth_date:
          type: string
          format: date
          example: "1965-07-31"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAuthorRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "New Author"
        bio:
          type: string
        birth_date:
          type: string
          format: date
    UpdateAuthorRequest:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        birth_date:
          type: string
          format: date

    # Pagination
    PageResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        meta:
          type: object
          properties:
            total:
              type: integer
              description: Total number of items
            page:
              type: integer
              description: Current page number
            limit:
              type: integer
              description: Items per page
            totalPages:
              type: integer
              description: Total number of pages
            hasNextPage:
              type: boolean
              description: Whether there is a next page
            hasPrevPage:
              type: boolean
              description: Whether there is a previous page

    # Errors
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: "Bad Request"
        error:
          type: string
          example: "Bad Request"